Add Kafka producer for student to join tournament
This commit introduces the ability for a student to join a tournament by sending a Kafka message. A new Kafka producer, StudentJoinTournamentKafkaProducer, has been added to the service. Upon successful tournament join, the producer sends a Kafka message with the details. Some comments were also simplified in the Score entity.