Update tournament service for student join/leave operations
The commit refines the process of a student joining a tournament and introduces the ability for a student to leave a tournament. This was mainly achieved by renaming StudentJoinDto to StudentJoinTournamentDto and adding StudentQuitTournamentDto. Matching changes were made in other service classes and a new kafka producer, StudentQuitTournamentKafkaProducer, has been added for when a student leaves a tournament.