Change return type of addEducatorToTournament method
The return type of the addEducatorToTournament method in the EducatorService interface, its implementation and use in the EducatorController has been changed from Educator to Tournament. This refactoring reflects that the method now returns the updated Tournament instead of the newly added Educator. In addition, logging via Slf4j has been added to EducatorController.