Refined security rules according to user roles
Security configuration in SecurityConfig.java has been updated to handle request authorization based on the role of the user; namely 'EDUCATOR' and 'STUDENT'. This is done by creating RequestMatchers for educator and student specific endpoints, as well as for common ones, in order to streamline and enhance access control in the system.