Add educator organizers to tournament
A ManyToMany relationship has been added to the Tournament entity for organizers. This allows educators to be added as organizers to a tournament. A new Kafka topic 'tournament.educator.join' and corresponding Kafka consumer 'EducatorJoinTournamentKafkaConsumer' has been created to accommodate these changes. Also, additional properties for Student and Tournament entities have been added, along with renaming 'educatorId' to 'creatorId' in the DTO.