Add build files and update application configurations and DTO
This commit introduces several build files for frontend, including manifest.json and robots.txt, implying the setup of the react app. It also updates the passwords in the configuration files of api-gateway and user-service. The rename and update in 'CreateBattleDto.java' suggests refactoring and enriching the data transfer object for clearer operation. Additionally, the new license file embraces third-party licences, likely due to the usage of external libraries/modules.