Add participants to TournamentDto
The class TournamentDto has been updated to include a new field 'participantsIds' to track the list of students participating in a tournament. Also, a relevant mapping method has also been added to the 'fromEntity' function to handle this new field. The relevant import statement for Student entity has been added as well.