"Improve user creation and tournament management functions"
This commit refines the user creation process and improves tournament handling. It incorporates changes including adding userId in the AuthenticationResponse, removing the password field in the User entity, and creating a new function to fetch all tournaments. Additionally, a description field for tournaments has been introduced and a more informative response message for successful logins is now implemented in the Authentication Controller.