Add TournamentRepository and related entities
Added the TournamentRepository interface with a method to find all tournaments in preparation with deadlines after a certain date. Also added the related entity classes Battle and Tournament. Two new services, BattleRoutine and TournamentRoutine, were created to potentially handle the tournament deadlines. Spring Data JPA is now included in the application dependencies and the application's connection to a MySQL database has been configured.