Add Dockerfile, update entities and tests, and create DDL script
Added a Dockerfile to containerize the tournament service. Made changes to various entities like Educator, Score, ScoreId, and Student to enhance the database schema implementation. Moreover, deleted the TournamentControllerTest due to changes in the service structure. A DDL script for database creation was added. Lastly, the ScoreId class was moved to a more appropriate package 'compositeKeys'.