Integrate Kafka messaging for Battle status changes
The BattleRoutine class was refactored to publish Battle status changes to a new Kafka topic. The status change messages are created as instances of the new ChangeBattleStatusDto class. Any exceptions during the JSON processing of these messages will cause a RuntimeException to be thrown. All of these changes also necessitated a move of the BattleRoutine class from the timeServices package to the service.timeServices package.