Implement email notifications for new tournaments and battles
An email notification feature has been added. Now, whenever a new tournament or battle is created, an automatic email notification is sent to the relevant parties. This update also includes the necessary changes in related classes, the addition of two new utility classes for getting usernames and emails, and an update in the KafkaProducer class.