Refactor services to improve security and tournament notifications
Refactored GitServiceImpl to improve security by using gitHubToken from a property source rather than hard-coding it. In StudentNotificationService, replaced inline gathering of students with the use of a TournamentParticipantsGetter to improve clarity and modularity. Added a controller method in TournamentController to get list of participant IDs, making it easier to fetch participant information for multiple services or instances.